Port to camunda 7.5.0-alpha2

+++ src/main/java/org/camunda/bpm/engine/cassandra/provider/serializer/JobDefinitionEntitySerializer.java
@@ -1,9 +1,9 @@
 /* Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  *      http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -34,7 +34,7 @@ public class JobDefinitionEntitySerializer implements CassandraSerializer<JobDef
     data.setString("type", entity.getJobType());
     data.setString("config", entity.getJobConfiguration());
     if(entity.getOverridingJobPriority()!=null){
-      data.setInt("priority", entity.getOverridingJobPriority());
+      data.setLong("priority", entity.getOverridingJobPriority());
     }
     else{
       data.setToNull("priority");
@@ -53,7 +53,7 @@ public class JobDefinitionEntitySerializer implements CassandraSerializer<JobDef
     entity.setJobType(data.getString("type"));
     entity.setJobConfiguration(data.getString("config"));
     if(!data.isNull("priority")){
-      entity.setJobPriority(data.getInt("priority"));
+      entity.setJobPriority(data.getLong("priority"));
     }
     entity.setSuspensionState(data.getInt("suspension_state"));
     entity.setRevision(data.getInt("revision"));
